GAE/requirements.txt
blinker==1.7.0
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.51.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
lxml==5.2.0
MarkupSafe==2.1.5
matplotlib==3.8.4
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==1.4.1
pandas-datareader==0.10.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
urllib3==2.0 
Werkzeug==3.0.2
yfinance==0.1.70
